refactor(core): 拆分 services/problems.ts 为 5 子模块 + barrel
原文件 1268 行单文件 → 1424 行 6 文件（多出来的是每个文件独立的 doc comment / imports）。

物理拆分（与 submissions 拆分模式 PR-3 同构）：
- problems-types.ts (131 行) — ProblemResponse / ProblemListResponse /
  AdminProblemList* 接口 + validateRuntimeConfig
- problems-list.ts (371 行) — listProblems / listAllProblems / getProblem /
  getProblemByTypeAndNumber + 私有 attachCategories / toProblemResponse
- problems-categories.ts (51 行) — syncProblemCategories
- problems-crud.ts (359 行) — createProblem / updateProblem / deleteProblem
- problems-export.ts (468 行) — buildExportPayload / importProblems
  + 私有 parseImportPayload / importOne / buildCategoryLookup
- problems.ts (44 行) — barrel re-export，4 处既有 import 路径不变

依赖关系（无循环）：
  problems-types.ts ← problems-list.ts ← problems-categories.ts
                                       ← problems-crud.ts ← problems-export.ts
  全部由 problems.ts barrel 重新对外暴露

验证：
- deno fmt / lint / check:types 全部通过
- deno task test:smoke 通过（GET /problems 走 listProblems → list.ts）
- tests/services/problems.test.ts 与 tests/routes/problems.test.ts 类型检查通过
- 已知预存在 bug：judge_image NOT NULL 约束未写入（reproduce 于 origin/main，
  非本次拆分引入，与 PR #69 同源，留待后续单独跟进）
